backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
async-timeout==5.0.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
kubernetes==33.1.0
limits==5.5.0
llama_cpp_python==0.3.16
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
overrides==7.7.0
packaging==25.0
pillow==11.3.0
posthog==6.5.0
protobuf==6.32.0
psycopg2-binary==2.9.9
pulsar-client==3.8.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyPDF2==3.0.1
PyPika==0.48.9
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.1.1
python-multipart==0.0.6
PyYAML==6.0.2
redis==5.0.1
regex==2025.7.34
requests==2.31.0
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.27.0
structlog==23.2.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0
